+++ test/sample-notification-transport-processing-summary-spec.js
@@ -4,7 +4,7 @@ var errorFormatter = testHelper.validationErrorFormatter;
 
 describe('Sample notification transport processing summary doc definition', function() {
   beforeEach(function() {
-    testHelper.init('build/sync-functions/test-sample-sync-function.js');
+    testHelper.initSyncFunction('build/sync-functions/test-sample-sync-function.js');
   });
 
   function verifyProcessingSummaryWritten(doc, oldDoc) {