+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - [#95](https://github.com/Kashoo/synctos/issues/95): Accept Date object for date/time constraint parameters
 - [#94](https://github.com/Kashoo/synctos/issues/94): Support dynamic validation constraint definitions
+- [#100](https://github.com/Kashoo/synctos/issues/100): Option to initialize test helper module with document definition file
 
 ## [1.8.0] - 2017-03-21
 ### Added